analytics-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
influxdb-client==1.36.0
pydantic==1.10.8
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.3.1
pytest-asyncio==0.21.0